requirements.txt
Django==2.2.4
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-markdownx==2.0.28
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
